+++ Basic/Strict-mode-ErrorRecord-formatting/Test-2.Format-List.ps1
@@ -1,4 +1,5 @@
 
+# v6-rc: fixed
 [PowerShell]::Create().AddScript({
 	Set-StrictMode -Version 2
 	try {